# swagger.yaml
openapi: "3.0.0"
info:
  title: Tshirt Store API
  description: Backend API for ecom site
  version: 1.1.0
  contact:
    email: nitinchoudharyn14@gmail.com
    url: "http://localhost/api/v1"

# servers:
#   - url: "https://localhost:4000/api/v1"
#     description: for local host - secure
#   - url: "http://localhost:4000/api/v1"
#     description: for local host - regular
servers:
  - url: "{protocol}://{url}/api/{version}"
    description: TShirtStore Swagger API 
    variables:
      version:
        enum:
          - v1
          - v2
        default: v1
      url:
        enum:
          - localhost:4000
          - tshirtstore.herokuapp.com
        default: localhost:4000
      protocol:
        enum:
          - http
          - https
        default: http
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    BearerAuth:
      type: http
      scheme: bearer
paths:  
...
  /products:
    get:
      tags:
        - Product Management
      summary: A simple get request will give you all available products
        
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  filteredProductCount:
                    type: number
                  totalcountProduct:
                    type: number
                  products:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                        name:
                          type: string
                        price: 
                          type: number
                        description: 
                          type: string
                        category: 
                          type: string
                        brand: 
                          type: string
                        stock:
                          type: number
                        ratings: 
                          type: number
                        numberOfReviews: 
                          type: number
                        user: 
                          type: string
                        createdAt:
                          type: string
                        photos: 
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: string
                              secure_url:
                                type: string
                              _id:
                                type: string
                        reviews: 
                          type: array
                          items:
                            type: object
                            properties:
                              _id:
                                type: string
                              comment:
                                type: string
                              user:
                                type: string
                              name:
                                type: string
                              rating:
                                type: number
        400:
          description: Bad request
        500:
          description: internal server error
  /product/{id}:
    get:
      tags:
        - Product Management
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  product:
                    type: object
                    properties:
                      _id:
                        type: string
                      name:
                        type: string
                      price: 
                        type: number
                      description: 
                        type: string
                      category: 
                        type: string
                      brand: 
                        type: string
                      stock:
                        type: number
                      ratings: 
                        type: number
                      numberOfReviews: 
                        type: number
                      user: 
                        type: string
                      createdAt:
                        type: string
                      photos: 
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            secure_url:
                              type: string
                            _id:
                              type: string
                      reviews: 
                        type: array
                        items:
                          type: object
                          properties:
                            _id:
                              type: string
                            comment:
                              type: string
                            user:
                              type: string
                            name:
                              type: string
                            rating:
                              type: number
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        500:
          description: internal server error
  /admin/products:
    get:
      tags:
        - Product Management
      summary: If user is admin, get all products
        
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  products:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                        name:
                          type: string
                        price: 
                          type: number
                        description: 
                          type: string
                        category: 
                          type: string
                        brand: 
                          type: string
                        stock:
                          type: number
                        ratings: 
                          type: number
                        numberOfReviews: 
                          type: number
                        user: 
                          type: string
                        createdAt:
                          type: string
                        photos: 
                          type: array
                          items:
                            type: object
                            properties:
                              id:
                                type: string
                              secure_url:
                                type: string
                              _id:
                                type: string
                        reviews: 
                          type: array
                          items:
                            type: object
                            properties:
                              _id:
                                type: string
                              comment:
                                type: string
                              user:
                                type: string
                              name:
                                type: string
                              rating:
                                type: number
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        500:
          description: internal server error
  /admin/product/add:
    post:
      tags:
        - Product Management
      summary: If admin, Add a new product
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
                - price
                - description
                - category
                - stock
                - brand
                - photos
              properties:
                name:
                  type: string
                  required: true
                price:
                  type: number
                  required: true
                description:
                  type: string
                  required: true
                category:
                  type: string
                  required: true
                  enum: [shortsleeves, longsleeves, sweatshirt, hoodies]
                stock:
                  type: number
                  required: true
                brand:
                  type: string
                  required: true
                photos:
                  in: formData
                  description: Upload photos
                  type: array
                  items:
                    type: file
                    format: binary
        
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  product:
                    type: object
                    properties:
                      _id:
                        type: string
                      name:
                        type: string
                      price: 
                        type: number
                      description: 
                        type: string
                      category: 
                        type: string
                      brand: 
                        type: string
                      stock:
                        type: number
                      ratings: 
                        type: number
                      numberOfReviews: 
                        type: number
                      user: 
                        type: string
                      createdAt:
                        type: string
                      photos: 
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            secure_url:
                              type: string
                            _id:
                              type: string
                      reviews: 
                        type: array
                        items:
                          type: object
                          properties:
                            _id:
                              type: string
                            comment:
                              type: string
                            user:
                              type: string
                            name:
                              type: string
                            rating:
                              type: number
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        500:
          description: internal server error
  /admin/product/{id}:
    put:
      tags:
        - Product Management
      summary: If admin, update the product. photos are optional to pass. Rest all body is set to be updated to what you pass here.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: number
                description:
                  type: number
                category:
                  type: string
                  enum: [shortsleeves, longsleeves, sweatshirt, hoodies]
                stock:
                  type: number
                brand:
                  type: string
                photos:
                  in: formData
                  description: The uploaded photos
                  type: file
                  format: binary
        
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  product:
                    type: object
                    properties:
                      _id:
                        type: string
                      name:
                        type: string
                      price: 
                        type: number
                      description: 
                        type: string
                      category: 
                        type: string
                      brand: 
                        type: string
                      stock:
                        type: number
                      ratings: 
                        type: number
                      numberOfReviews: 
                        type: number
                      user: 
                        type: string
                      createdAt:
                        type: string
                      photos: 
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            secure_url:
                              type: string
                            _id:
                              type: string
                      reviews: 
                        type: array
                        items:
                          type: object
                          properties:
                            _id:
                              type: string
                            comment:
                              type: string
                            user:
                              type: string
                            name:
                              type: string
                            rating:
                              type: number
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        500:
          description: internal server error
    delete:
      tags:
        - Product Management
      summary: If admin, pass the id of product and product will be removed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
                
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        500:
          description: internal server error
  /product/reviews:
      get:
        tags:
          - Product Management
        summary: A simple get request will give you all reviews for given products
        parameters:
        - name: id
          in: query
          description: id of the product
          required: true
          schema:
            type: string
          
        responses:
          200:
            description: All good success
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    success:
                      type: boolean
                    Overall_Ratings:
                      type: number
                    Total_number_of_reviews:
                      type: number
                    reviews:
                      type: array
                      items:
                        type: object
                        properties:
                          user:
                            type: string
                          name:
                            type: string
                          rating: 
                            type: number
                          comment: 
                            type: string
                          _id: 
                            type: string
                          
          400:
            description: Bad request
            content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
          500:
            description: internal server error
  
      put:
        tags:
          - Product Management
        summary: A logged in user can post a review on any product. If review is already posted, it will just update the existing review
        
        requestBody:
          content:
            application/json:
              schema:
                type: object
                required:
                  - rating
                  - comment
                  - productID
                properties:
                  rating:
                    type: number
                    required: true
                  comment:
                    type: string
                    required: true
                  productID:
                    type: string
                    required: true
        responses:
          200:
            description: All good success
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    success:
                      type: boolean
                    message:
                      type: object
                      properties:
                        user:
                          type: string
                        name:
                          type: string
                        rating: 
                          type: number
                        comment: 
                          type: string
                          
          400:
            description: Bad request
            content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
          500:
            description: internal server error

      delete:
        tags:
          - Product Management
        summary: Delete the review of logged in user on given product id.
        parameters:
        - name: productID
          in: query
          required: true
          schema:
            type: string
        
        responses:
          200:
            description: All good success
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    success:
                      type: boolean
                    message:
                      type: string
                          
          400:
            description: Bad request
            content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
          500:
            description: internal server error


# order routes
...